fix: correct all clippy uninlined_format_args lints
- Fix format strings in blockmeta_client.rs to use inline variables
- Fix format strings in main.rs CLI implementation
- Fix unnecessary cast in serialize.rs test
- Ensure CI passes with --all-targets clippy check

This resolves all remaining clippy warnings that were failing CI.

+++ crates/encoding/src/serialize.rs
@@ -213,7 +213,7 @@ mod tests {
     fn encode_i64() {
         for (unsigned, signed) in ZIGZAG_TESTS.iter() {
             let mut buf_u64 = Vec::new();
-            serialize_u64(*unsigned as u64, &mut buf_u64);
+            serialize_u64(*unsigned, &mut buf_u64);
 
             let mut buf_i64 = Vec::new();
             serialize_i64(*signed, &mut buf_i64);
